Switch to use FastAPI
