--- app.py
@@ -1,6 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-# Import routes
-import pyris.routes.llm
