+++ requirements.txt
@@ -1,7 +1,6 @@
-Flask==2.3.2
+fastapi==0.95.1
 guidance==0.0.62
 pydantic==1.10.7
-Flask-Pydantic==0.11.0
 python-dotenv==1.0.0
 pytest==7.3.1
 pytest-mock==3.10.0
