+++ tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from app.main import app
+from fastapi.testclient import TestClient
+
+
+@pytest.fixture(scope="session")
+def test_client():
+    with TestClient(app) as c:
+        yield c
